--- a/src/hooks/useBalance.ts
+++ b/src/hooks/useBalance.ts
@@ -19,12 +19,12 @@ import { VoidFn } from '@polkadot/api/types';
 import { Balance } from '@polkadot/types/interfaces';
 import { formatBalance } from '@polkadot/util';
 import BN from 'bn.js';
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 
 import { MessageActionsCreators } from '../actions/messageActions';
 import { useUpdateMessageContext } from '../contexts/MessageContext';
 import logger from '../util/logger';
-import { useIsMounted } from './useIsMounted';
+import { useMountedState } from './useMountedState';
 
 type State = {
   chainTokens: string;
@@ -42,24 +42,21 @@ const initValues = {
 
 const useBalance = (api: ApiPromise, address: string, providedSi: boolean = false): State => {
   const { dispatchMessage } = useUpdateMessageContext();
-  const [state, setState] = useState<State>(initValues);
-
-  const isMounted = useIsMounted();
+  const [state, setState] = useMountedState<State>(initValues);
 
   useEffect((): (() => void) => {
     const getBalance = async (api: ApiPromise, address: string, setState: any): Promise<VoidFn> => {
       try {
         const u = await api.query.system.account(address, ({ data }): void => {
-          isMounted() &&
-            setState({
-              chainTokens: data.free.registry.chainTokens[0],
-              formattedBalance: formatBalance(data.free, {
-                decimals: api.registry.chainDecimals[0],
-                forceUnit: '-',
-                withSi: providedSi
-              }),
-              free: data.free
-            });
+          setState({
+            chainTokens: data.free.registry.chainTokens[0],
+            formattedBalance: formatBalance(data.free, {
+              decimals: api.registry.chainDecimals[0],
+              forceUnit: '-',
+              withSi: providedSi
+            }),
+            free: data.free
+          });
         });
         return Promise.resolve(u);
       } catch (e) {
@@ -77,7 +74,7 @@ const useBalance = (api: ApiPromise, address: string, providedSi: boolean = fals
     return async (): Promise<void> => {
       unsubscribe && (await unsubscribe)();
     };
-  }, [address, providedSi, dispatchMessage, api, setState, isMounted]);
+  }, [address, providedSi, dispatchMessage, api, setState]);
 
   return state as State;
 };