Memory leak was coming from useMountedState.
There is something in that custom hook that was triggering loops (irrelevant completely to how the useEffect is written - chain   promise or async/await).
Will investigate in different PR how to make the useMountedState usable instead of using isMounted && setState - but for now I would like to fix the memory leak.