@tomusdrw 2021-05-06
> and whenever the new value arrives - React could not understand that the object changed or not - so it was forcing a re-render

Can you elaborate how re-rendering is connected with the leak? Cause I don't see immediately how fixing re-rendering fixes the leak as well - I guess it's just made less severe?