@hbulgarini 2021-05-06
> > and whenever the new value arrives - React could not understand that the object changed or not - so it was forcing a re-render
> 
> Can you elaborate how re-rendering is connected with the leak? Cause I don't see immediately how fixing re-rendering fixes the leak as well - I guess it's just made less severe?

i think this PR solves more the re-rendering loop rather than the memory leak .. actually comparing the performance of my PR #136 originated from branch originated from a master version previous to the useMountedState, the performance was not so bad ... actually it was perfectly usable .
but this PR solves the infinite loop re-rendering and in my opinion this is why i would go ahead with the merge of this PR as it is since it solves an issue with is deployed now