@wirednkod 2021-05-07
> > and whenever the new value arrives - React could not understand that the object changed or not - so it was forcing a re-render
> 
> Can you elaborate how re-rendering is connected with the leak? Cause I don't see immediately how fixing re-rendering fixes the leak as well - I guess it's just made less severe?

Just to make things clear I will try to place a bit of what happened and what I mean by "leak".
I wanted to identify the Error that was appearing with description the "cannot update state when component is not mounted". In order to do so - I introduced the useIsMounted hook that eventually was replaced by the useMountedState hook. After that I checked the request from @tomusdrw concerning a potential "leak" (I'll explain why the "") that may or may not exist concerning the subscriptions for the useBalance. 
Once I tried to identify that I realised that the "leak" is actually a re-rendering issue that forces useBalance to run non-stop (thus not a leak but a "leak"). I thought initially that the problem was identified from the Promise.resolve/reject that were not run in time and alter the code to .then pattern. Unfortunately the case was lying with the useMountedState hook - that was forcing re-render of the components using it - forcing re-use of the useBalance and thus continuous subscriptions/un-subscriptions that was calling api non-stop and overloading memory/CPU usage. This can be seen in the screenshot below 

![Screenshot 2021-05-06 at 8 17 51 PM](https://user-images.githubusercontent.com/5408605/117410191-c55aba00-af1a-11eb-8e03-79b576140aa4.png)

From the Summary at the bottom it was identified that "leak" was due to what is marked as Scripting _**(wavy pattern of JS Heap - blue line).**_. Meaning that JavaScript code was triggering expensive calls/recalculation/re-rendering that caused UI to lag. This was one of the major causes of poor performance or as I named it "leak".  After replacing the useMountedState to useState and revert code to use the `isMounted() && useState` the outcome can be seen  on following screenshot:

![Screenshot 2021-05-06 at 8 39 28 PM](https://user-images.githubusercontent.com/5408605/117410857-9db82180-af1b-11eb-9b74-c5e6b2e564ea.png)

The Scripting actions are reduced from 4,78 secs to 1,16 secs and specific action's loading time from 5,066 secs to 1,521 secs.

So the "leak" as I mentioned before was not the suspicion of subscriptions initialising but never close - but the continuous/non-stop re-initialisation of subscriptions that were though closing.

With the solution given here - the lifecycle of React is now controlled and is triggering only when it actually needs to (state must be updated).
In order to make the useMountedState to work - a deep checking on incoming value (especially on objects/arrays) need to take place to ensure if re-set of state is needed or not. I intend to do that in a future PR.